BE - Fix Langflow response

+++ BE/FPT_EduTrack/FPT_EduTrack.BusinessLayer/DTOs/Response/GradingResponse.cs
@@ -4,7 +4,7 @@
     {
         public string? SessionId { get; set; }
         public string? Timestamp { get; set; }
-        public int OverallBand { get; set; }
+        public double OverallBand { get; set; }
         public Dictionary<string, Criterion> Criteria { get; set; } = new();
         public string Justification { get; set; } = "";
         //public string FullEvaluation { get; set; } = "";
